refactor: move VideosService into Video module
