+++ src/app/video/pipes/filter-videos-by-title.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { Video } from '../../video-data/models/video';
+import { Video } from '../models/video';
 
 @Pipe({
   name: 'filterVideosByTitle',
