+++ src/app/video/video-grid/video-grid.component.ts
@@ -1,7 +1,7 @@
-import { VideosService } from '../../video-data/services/videos.service';
+import { VideosService } from 'src/app/video/services/videos.service';
 import { Component, OnInit } from '@angular/core';
-import { Video } from '../../video-data/models/video';
-import { VideoFilter } from '../../video-data/models/video-filter';
+import { Video } from '../models/video';
+import { VideoFilter } from '../models/video-filter';
 
 @Component({
   selector: 'app-video-grid',
